ENH(TST): Skip a test when cannot remove curent directory

+++ datalad/support/tests/test_path.py
@@ -7,11 +7,13 @@
 #
 # ## ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ### ##
 
+import os
 from ..path import (
     abspath,
     curdir,
     robust_abspath,
 )
+from ...dochelpers import exc_str
 from ...utils import (
     chpwd,
     rmtree,
@@ -20,6 +22,7 @@ from ...tests.utils import (
     assert_raises,
     eq_,
     with_tempfile,
+    SkipTest,
 )
 
 
@@ -27,6 +30,14 @@ from ...tests.utils import (
 def test_robust_abspath(tdir):
     with chpwd(tdir):
         eq_(robust_abspath(curdir), tdir)
-        rmtree(tdir)
+        try:
+            if os.environ.get('DATALAD_ASSERT_NO_OPEN_FILES'):
+                raise Exception("cannot test under such pressure")
+            rmtree(tdir)
+        except Exception as exc:
+            # probably windows or above exception
+            raise SkipTest(
+                "Cannot test in current environment: %s" % exc_str(exc))
+
         assert_raises(OSError, abspath, curdir)
         eq_(robust_abspath(curdir), tdir)
